+++ NEWS.md
@@ -2,6 +2,11 @@
 
 * Respect custom content type for `req_body_json()` (@mgirlich, #189).
 
+* `req_oauth_refresh()` now respects the `refresh_token` for caching (@mgirlich, #178).
+
+* `req_url_path()` and `req_url_path_append()` can now handle `NULL` or empty
+  `...` and the elements of `...` can also have length > 1 (@mgirlich, #177).
+
 * `req_perform()` now has an `error_call` argument and communicates more clearly
   where the error occurred (@mgirlich, #187).